Change HTTP handler API for clearer mem management
With the old API and usage where the export function of a GCF is set to
`slacktion.getExportFunction()` the `slacktion` variable goes out of
scope and the instance is being kept alive by the `this` reference in
the returned function.

This makes for a more confusing and harder to reason about memory
management which can often lead to problems down the road. It's better
to be more explicit about who is owning the strong reference.

Changing this also mean we can remove a check for the type of Slacktion
(HTTP or RTM since the API doesn't make sense to call without HTTP
request and response objects).