Use CSP channels for syncing reply messages
This is a solution to #8 where multiple messages can get out of order.

The way it works is that for each incoming message, a reply synchronizer is
spawned with a reply channel it listens to, the replyFn will simply put
a reply on the channel and the reply synchronizer will read one reply at
a time from the channel and send it back to Slack.